requirements.txt
flask-diamond>=0.5.0
cryptography
passlib>=1.7.0
ldap3
pyasn1
pyasn1_modules
Flask-DebugToolbar>=0.10.0
Flask-Bootstrap
Flask-Babel
Flask-QRcode
pyopenssl

# Due to https://github.com/flask-admin/flask-admin/issues/1588
sqlalchemy==1.2.0b3